Refactored to match new master setup
