+++ client/components/Channel/Reaction.js
